- use int instead float for cordiantion

+++ master_src/MainActivity.java
@@ -101,7 +101,7 @@ public class MainActivity extends Activity {
 			s_game_state = gameState.STATE_INIT;
 			s_init_state = true;
 			#ifdef USE_BG_BUFFER
-			s_touch = new myTouch(SCREEN_WIDTH/SCREEN_WIDTH_PHONE, SCREEN_HEIGHT/SCREEN_HEIGHT_PHONE);
+			s_touch = new myTouch((float)SCREEN_WIDTH/SCREEN_WIDTH_PHONE, (float)SCREEN_HEIGHT/SCREEN_HEIGHT_PHONE);
 			s_bg_buffer = Bitmap.createBitmap((int)SCREEN_WIDTH, (int)SCREEN_HEIGHT, Bitmap.Config.ARGB_8888);
 			s_canvas = new Canvas(s_bg_buffer);
 			#else
